pyproject.toml
[tool.poetry]
name = "scouting-tools"
description = "Team 3256's scouting utilities"
authors = []
version = "0.1.0"

readme = "README.md"
license = "GPL-3.0-or-later"

classifiers = [
    # Get the list of trove classifiers here: https://pypi.org/classifiers/
    "Programming Language :: Python :: Implementation :: CPython",
    "Typing :: Typed",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Natural Language :: English",
]
documentation = "https://scoutingtools.readthedocs.io/en/latest/index.html"
homepage = "https://github.com/ThatXliner/scouting-tools"
# keywords = ["awesome", "project"]  # Maximum of 5 keywords

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
diskcache = "^5.6.3"

[tool.poetry.scripts]
scouting_tools = "scouting_tools.__main__:main"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
hypothesis = "^6.68.2"
mypy = "^1.0.1"
pytest = "^7.2.1"
pytest-clarity = "^1.0.1"
pytest-cov = "^4.0.0"
ruff = "^0.2.2"
shed = "^0.10.9"

Sphinx = "^6.1.3"
toml = "^0.10.2"

[tool.poe.tasks]
# Code linting
mypy = { cmd = "mypy scouting_tools --strict", help = "Run MyPy on codebase" }
ruff = { cmd = "ruff check scouting_tools", help = "Run Ruff on codebase" }
check_black = { cmd = "black scouting_tools --check" }
check_imports = { cmd = "ruff check scouting_tools --select I" }
style = ["check_black", "check_imports"]
codebase = ["ruff", "mypy"]
[tool.poe.tasks.docs]
cmd = "sphinx-build docs build"
help = "Build documentation"

[tool.poe.tasks.format]
cmd = "shed"
help = "Format code"

[tool.poe.tasks.fix-ruff]
cmd = "ruff scouting_tools --fix"
help = "Ruff autofix"

[tool.poe.tasks.lint]
sequence = ["style", "codebase"]
help = "Lint codebase"

[tool.poe.tasks.test]
cmd = "pytest -vvv --cov=scouting_tools"
help = "Simply run test suite"

[tool.poe.tasks.ci]
cmd = "pytest -vvv --cov=scouting_tools --cov-report=xml"
help = "This workflow is for Github Actions"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Same as Black.
line-length = 88
target-version = "py38"
[tool.ruff.lint]
select = ["ALL"]
ignore = [
    # "D",    # "No docs"
    "T20", # "Don't use print or pprint"
    # "ANN",  # Type annotation errors (or the lack of it)
    "ANN101", # The type annotation for `self` is inferred
    "FBT",    # I actually don't know why this exists
    # and it seems useless so ignore it
    # Fix doc rule conflicts
    "D203",
    "D213",
]
[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "D"]
"docs/conf.py" = ["INP001", "A001"]
